requirements.txt
python-dotenv==1.0.0
pandas==2.1.0
jinja2==3.1.2
flask==3.0.2
flask-wtf==1.2.1
gunicorn==21.2.0
whitenoise==6.6.0
sendgrid==6.11.0
flask-cors==4.0.0 